+++ cmd/lfgw/lf.go
@@ -34,22 +34,38 @@ func (app *application) isFakePositiveRegexp(filter metricsql.LabelFilter) bool 
 	return false
 }
 
-// shouldBeModified helps to understand whether the original label filter has to be mofified. It returns true if [the list of original filters contains either a fake positive regexp (no special symbols, e.g. namespace=~"kube-system") or a non-regexp filter] and [newFilter is a matching positive regexp]. Target label is taken from the newFilter.
-func (app *application) shouldBeModified(filters []metricsql.LabelFilter, newFilter metricsql.LabelFilter) bool {
+// TODO: update description?
+// shouldBeModified helps to understand whether the original label filters have to be modified. The function returns false if any of the original filters do not match expectations described further. It returns false if [the list of original filters contains either a fake positive regexp (no special symbols, e.g. namespace=~"kube-system") or a non-regexp filter] and [newFilter is a matching positive regexp]. Also, if the original and the new filter are equal. Target label is taken from the newFilter.
+func (app *application) shouldNotBeModified(filters []metricsql.LabelFilter, newFilter metricsql.LabelFilter) bool {
+	seen := 0
+	seenUnmodified := 0
+
 	for _, filter := range filters {
 		if filter.Label == newFilter.Label && newFilter.IsRegexp && !newFilter.IsNegative {
+			seen++
+
+			// Target: non-regexps or fake regexps
 			if !filter.IsRegexp || app.isFakePositiveRegexp(filter) {
 				// Prometheus treats all regexp queries as anchored, whereas our raw regexp doesn't have them. So, we should take anchored values.
 				re, err := metricsql.CompileRegexpAnchored(newFilter.Value)
 				// There shouldn't be any errors, though, just in case, better to skip deduplication
 				if err == nil && re.MatchString(filter.Value) {
-					return false
+					seenUnmodified++
+					continue
 				}
 			}
+
+			// Target: both are positive regexps with the same value
+			if app.equalLabelFilters(filter, newFilter) {
+				seenUnmodified++
+				continue
+			}
+
+			// TODO: check if matches any subfilter
 		}
 	}
 
-	return true
+	return seen > 0 && seen == seenUnmodified
 }
 
 // appendOrMergeRegexpLF appends label filter or merges its value in case it's a regexp with the same name and of the same type (negative / positive).
@@ -91,7 +107,7 @@ func (app *application) modifyMetricExpr(expr metricsql.Expr, newFilter metricsq
 	modifyLabelFilter := func(expr metricsql.Expr) {
 		if me, ok := expr.(*metricsql.MetricExpr); ok {
 			if newFilter.IsRegexp {
-				if app.shouldBeModified(me.LabelFilters, newFilter) || !app.EnableDeduplication {
+				if !app.shouldNotBeModified(me.LabelFilters, newFilter) || !app.EnableDeduplication {
 					me.LabelFilters = app.appendOrMergeRegexpLF(me.LabelFilters, newFilter)
 				}
 			} else {
@@ -126,6 +142,11 @@ func (app *application) optimizeMetricExpr(expr metricsql.Expr) metricsql.Expr {
 // equalExpr says whether two expressions are equal.
 func (app *application) equalExpr(expr1 metricsql.Expr, expr2 metricsql.Expr) bool {
 	return string(expr1.AppendString(nil)) == string(expr2.AppendString(nil))
+}
+
+// equalLabelFilters says whether two labelfilters are equal.
+func (app *application) equalLabelFilters(lf1, lf2 metricsql.LabelFilter) bool {
+	return string(lf1.AppendString(nil)) == string(lf2.AppendString(nil))
 }
 
 // prepareQueryParams rewrites GET/POST "query" and "match" parameters to filter out metrics.